Commentary fixes
